+++ eslint.config.js
@@ -1,0 +1,50 @@
+// @ts-check
+
+import react from "eslint-plugin-react";
+import reactRefresh from "eslint-plugin-react-refresh";
+import globals from "globals";
+import tseslint from "typescript-eslint";
+
+import path from "path";
+import { fileURLToPath } from "url";
+
+// mimic CommonJS variables -- not needed if using CommonJS
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+export default [
+  {
+    files: ["lib/**/*.{ts,tsx}"],
+    // ignores: ["dist", "docs", "example", "**/eslint.config.js"],
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+        ...globals.es2020,
+      },
+      parser: tseslint.parser,
+      parserOptions: {
+        ecmaVersion: "latest",
+        sourceType: "module",
+        project: ["./tsconfig.json", "./tsconfig.node.json"],
+        tsconfigRootDir: __dirname,
+      },
+    },
+    plugins: {
+      "@typescript-eslint": tseslint.plugin,
+      "react-refresh": reactRefresh,
+      react: react,
+    },
+    rules: {
+      "react-refresh/only-export-components": [
+        "warn",
+        { allowConstantExport: true },
+      ],
+      "@typescript-eslint/no-unsafe-return": "error",
+    },
+  },
+];
+
+/* migrate documents
+ref1: https://eslint.org/blog/2022/08/new-config-system-part-2
+ref2: https://allalmohamedlamine.medium.com/eslint-flat-config-and-new-system-an-ultimate-deep-dive-2023-46aa151cbf2b
+*/